Restructure e2e tests into logical layout (#2064)
