+++ .github/workflows/e2e-fork.yml
@@ -24,7 +24,7 @@ jobs:
       - id: set-matrix
         run: echo "::set-output name=matrix::$(go run cmd/build_test_matrix/main.go)"
         env:
-          TEST_EXCLUSIONS: "TestInterchainAccountsTestSuite"
+          TEST_EXCLUSIONS: "TestInterchainAccountsTestSuite,TestIncentivizedInterchainAccountsTestSuite"
 
   e2e:
     env:
@@ -51,4 +51,4 @@ jobs:
           RLY_TAG: "v2.0.0-rc2"
         run: |
           cd e2e
-          make e2e-test suite=${{ matrix.suite }} test=${{ matrix.test }}
+          make e2e-test entrypoint=${{ matrix.entrypoint }} test=${{ matrix.test }}
