+++ .github/workflows/e2e-test-workflow-call.yml
@@ -129,4 +129,4 @@ jobs:
       - name: Run e2e Test
         run: |
           cd e2e
-          make e2e-test suite=${{ matrix.suite }} test=${{ matrix.test }}
+          make e2e-test entrypoint=${{ matrix.entrypoint }} test=${{ matrix.test }}
