+++ e2e/scripts/test-matricies/main/test-matrix.json
@@ -30,7 +30,7 @@
     ]
   },
   {
-    "test-entry-point": "TestFeeMiddlewareTestSuite",
+    "test-entry-point": "TestIncentivizedTransferTestSuite",
     "chain-binary": "simd",
     "tests": [
       {
@@ -55,6 +55,20 @@
         "chain-a-tag": "v0.1.3",
         "chain-b-tag": "v0.3.2"
       },
+      {
+        "chain-a-tag": "v0.2.2",
+        "chain-b-tag": "v0.3.2"
+      },
+      {
+        "chain-a-tag": "v0.3.2",
+        "chain-b-tag": "v0.2.2"
+      }
+    ]
+  },
+  {
+    "test-entry-point": "TestIncentivizedInterchainAccountsTestSuite",
+    "chain-binary": "icad",
+    "tests": [
       {
         "chain-a-tag": "v0.2.2",
         "chain-b-tag": "v0.3.2"
