+++ e2e/tests/interchain_accounts/incentivized_test.go
@@ -1,243 +1,31 @@
-package e2e
+package interchain_accounts
 
 import (
 	"context"
 	"testing"
 
-	ibctest "github.com/strangelove-ventures/ibctest"
-	"github.com/strangelove-ventures/ibctest/chain/cosmos"
-	"github.com/strangelove-ventures/ibctest/ibc"
-	"github.com/strangelove-ventures/ibctest/test"
-	"github.com/stretchr/testify/suite"
-	"golang.org/x/mod/semver"
-
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
 	intertxtypes "github.com/cosmos/interchain-accounts/x/inter-tx/types"
+	ibctest "github.com/strangelove-ventures/ibctest"
+	"github.com/strangelove-ventures/ibctest/ibc"
+	"github.com/strangelove-ventures/ibctest/test"
+	"github.com/stretchr/testify/suite"
 
-	"github.com/cosmos/ibc-go/e2e/testconfig"
-	"github.com/cosmos/ibc-go/e2e/testsuite"
 	"github.com/cosmos/ibc-go/e2e/testvalues"
-
-	icatypes "github.com/cosmos/ibc-go/v5/modules/apps/27-interchain-accounts/types"
 	feetypes "github.com/cosmos/ibc-go/v5/modules/apps/29-fee/types"
 	ibctesting "github.com/cosmos/ibc-go/v5/testing"
 )
 
-func TestInterchainAccountsTestSuite(t *testing.T) {
-	suite.Run(t, new(InterchainAccountsTestSuite))
+func TestIncentivizedInterchainAccountsTestSuite(t *testing.T) {
+	suite.Run(t, new(IncentivizedInterchainAccountsTestSuite))
 }
 
-type InterchainAccountsTestSuite struct {
-	testsuite.E2ETestSuite
+type IncentivizedInterchainAccountsTestSuite struct {
+	InterchainAccountsTestSuite
 }
 
-// RegisterInterchainAccount will attempt to register an interchain account on the counterparty chain.
-func (s *InterchainAccountsTestSuite) RegisterInterchainAccount(ctx context.Context, chain *cosmos.CosmosChain, user *ibc.Wallet, msgRegisterAccount *intertxtypes.MsgRegisterAccount) error {
-	txResp, err := s.BroadcastMessages(ctx, chain, user, msgRegisterAccount)
-	s.Require().NoError(err)
-	s.AssertValidTxResponse(txResp)
-	return err
-}
-
-// RegisterCounterPartyPayee broadcasts a MsgRegisterCounterpartyPayee message.
-func (s *InterchainAccountsTestSuite) RegisterCounterPartyPayee(ctx context.Context, chain *cosmos.CosmosChain,
-	user *ibc.Wallet, portID, channelID, relayerAddr, counterpartyPayeeAddr string) (sdk.TxResponse, error) {
-	msg := feetypes.NewMsgRegisterCounterpartyPayee(portID, channelID, relayerAddr, counterpartyPayeeAddr)
-	return s.BroadcastMessages(ctx, chain, user, msg)
-}
-
-// getICAVersion returns the version which should be used in the MsgRegisterAccount broadcast from the
-// controller chain.
-func getICAVersion(chainAVersion, chainBVersion string) string {
-	chainBIsGreaterThanChainA := semver.Compare(chainAVersion, chainBVersion) == -1
-	if chainBIsGreaterThanChainA {
-		// allow version to be specified by the controller chain
-		return ""
-	}
-	// explicitly set the version string because the host chain might not yet support incentivized channels.
-	return icatypes.NewDefaultMetadataString(ibctesting.FirstConnectionID, ibctesting.FirstConnectionID)
-}
-
-func (s *InterchainAccountsTestSuite) TestMsgSubmitTx_SuccessfulTransfer() {
-	t := s.T()
-	ctx := context.TODO()
-
-	// setup relayers and connection-0 between two chains
-	// channel-0 is a transfer channel but it will not be used in this test case
-	relayer, _ := s.SetupChainsRelayerAndChannel(ctx)
-	chainA, chainB := s.GetChains()
-
-	// setup 2 accounts: controller account on chain A, a second chain B account.
-	// host account will be created when the ICA is registered
-	controllerAccount := s.CreateUserOnChainA(ctx, testvalues.StartingTokenAmount)
-	chainBAccount := s.CreateUserOnChainB(ctx, testvalues.StartingTokenAmount)
-	var hostAccount string
-
-	t.Run("register interchain account", func(t *testing.T) {
-		version := getICAVersion(testconfig.GetChainATag(), testconfig.GetChainBTag())
-		msgRegisterAccount := intertxtypes.NewMsgRegisterAccount(controllerAccount.Bech32Address(chainA.Config().Bech32Prefix), ibctesting.FirstConnectionID, version)
-		err := s.RegisterInterchainAccount(ctx, chainA, controllerAccount, msgRegisterAccount)
-		s.Require().NoError(err)
-	})
-
-	t.Run("start relayer", func(t *testing.T) {
-		s.StartRelayer(relayer)
-	})
-
-	t.Run("verify interchain account", func(t *testing.T) {
-		var err error
-		hostAccount, err = s.QueryInterchainAccount(ctx, chainA, controllerAccount.Bech32Address(chainA.Config().Bech32Prefix), ibctesting.FirstConnectionID)
-		s.Require().NoError(err)
-		s.Require().NotZero(len(hostAccount))
-
-		channels, err := relayer.GetChannels(ctx, s.GetRelayerExecReporter(), chainA.Config().ChainID)
-		s.Require().NoError(err)
-		s.Require().Equal(len(channels), 2)
-	})
-
-	t.Run("interchain account executes a bank transfer on behalf of the corresponding owner account", func(t *testing.T) {
-
-		t.Run("fund interchain account wallet", func(t *testing.T) {
-			// fund the host account account so it has some $$ to send
-			err := chainB.SendFunds(ctx, ibctest.FaucetAccountKeyName, ibc.WalletAmount{
-				Address: hostAccount,
-				Amount:  testvalues.StartingTokenAmount,
-				Denom:   chainB.Config().Denom,
-			})
-			s.Require().NoError(err)
-		})
-
-		t.Run("broadcast MsgSubmitTx", func(t *testing.T) {
-			// assemble bank transfer message from host account to user account on host chain
-			msgSend := &banktypes.MsgSend{
-				FromAddress: hostAccount,
-				ToAddress:   chainBAccount.Bech32Address(chainB.Config().Bech32Prefix),
-				Amount:      sdk.NewCoins(testvalues.DefaultTransferAmount(chainB.Config().Denom)),
-			}
-
-			// assemble submitMessage tx for intertx
-			msgSubmitTx, err := intertxtypes.NewMsgSubmitTx(
-				msgSend,
-				ibctesting.FirstConnectionID,
-				controllerAccount.Bech32Address(chainA.Config().Bech32Prefix),
-			)
-			s.Require().NoError(err)
-
-			// broadcast submitMessage tx from controller account on chain A
-			// this message should trigger the sending of an ICA packet over channel-1 (channel created between controller and host)
-			// this ICA packet contains the assembled bank transfer message from above, which will be executed by the host account on the host chain.
-			resp, err := s.BroadcastMessages(
-				ctx,
-				chainA,
-				controllerAccount,
-				msgSubmitTx,
-			)
-
-			s.AssertValidTxResponse(resp)
-			s.Require().NoError(err)
-
-			s.Require().NoError(test.WaitForBlocks(ctx, 10, chainA, chainB))
-		})
-
-		t.Run("verify tokens transferred", func(t *testing.T) {
-			balance, err := chainB.GetBalance(ctx, chainBAccount.Bech32Address(chainB.Config().Bech32Prefix), chainB.Config().Denom)
-			s.Require().NoError(err)
-
-			_, err = chainB.GetBalance(ctx, hostAccount, chainB.Config().Denom)
-			s.Require().NoError(err)
-
-			expected := testvalues.IBCTransferAmount + testvalues.StartingTokenAmount
-			s.Require().Equal(expected, balance)
-		})
-	})
-}
-
-func (s *InterchainAccountsTestSuite) TestMsgSubmitTx_FailedTransfer_InsufficientFunds() {
-	t := s.T()
-	ctx := context.TODO()
-
-	// setup relayers and connection-0 between two chains
-	// channel-0 is a transfer channel but it will not be used in this test case
-	relayer, _ := s.SetupChainsRelayerAndChannel(ctx)
-	chainA, chainB := s.GetChains()
-
-	// setup 2 accounts: controller account on chain A, a second chain B account.
-	// host account will be created when the ICA is registered
-	controllerAccount := s.CreateUserOnChainA(ctx, testvalues.StartingTokenAmount)
-	chainBAccount := s.CreateUserOnChainB(ctx, testvalues.StartingTokenAmount)
-	var hostAccount string
-
-	t.Run("register interchain account", func(t *testing.T) {
-		version := getICAVersion(testconfig.GetChainATag(), testconfig.GetChainBTag())
-		msgRegisterAccount := intertxtypes.NewMsgRegisterAccount(controllerAccount.Bech32Address(chainA.Config().Bech32Prefix), ibctesting.FirstConnectionID, version)
-		err := s.RegisterInterchainAccount(ctx, chainA, controllerAccount, msgRegisterAccount)
-		s.Require().NoError(err)
-	})
-
-	t.Run("start relayer", func(t *testing.T) {
-		s.StartRelayer(relayer)
-	})
-
-	t.Run("verify interchain account", func(t *testing.T) {
-		var err error
-		hostAccount, err = s.QueryInterchainAccount(ctx, chainA, controllerAccount.Bech32Address(chainA.Config().Bech32Prefix), ibctesting.FirstConnectionID)
-		s.Require().NoError(err)
-		s.Require().NotZero(len(hostAccount))
-
-		channels, err := relayer.GetChannels(ctx, s.GetRelayerExecReporter(), chainA.Config().ChainID)
-		s.Require().NoError(err)
-		s.Require().Equal(len(channels), 2)
-	})
-
-	t.Run("fail to execute bank transfer over ICA", func(t *testing.T) {
-		t.Run("verify empty host wallet", func(t *testing.T) {
-			hostAccountBalance, err := chainB.GetBalance(ctx, hostAccount, chainB.Config().Denom)
-			s.Require().NoError(err)
-			s.Require().Zero(hostAccountBalance)
-		})
-
-		t.Run("broadcast MsgSubmitTx", func(t *testing.T) {
-			// assemble bank transfer message from host account to user account on host chain
-			transferMsg := &banktypes.MsgSend{
-				FromAddress: hostAccount,
-				ToAddress:   chainBAccount.Bech32Address(chainB.Config().Bech32Prefix),
-				Amount:      sdk.NewCoins(testvalues.DefaultTransferAmount(chainB.Config().Denom)),
-			}
-
-			// assemble submitMessage tx for intertx
-			submitMsg, err := intertxtypes.NewMsgSubmitTx(
-				transferMsg,
-				ibctesting.FirstConnectionID,
-				controllerAccount.Bech32Address(chainA.Config().Bech32Prefix),
-			)
-			s.Require().NoError(err)
-
-			// broadcast submitMessage tx from controller account on chain A
-			// this message should trigger the sending of an ICA packet over channel-1 (channel created between controller and host)
-			// this ICA packet contains the assembled bank transfer message from above, which will be executed by the host account on the host chain.
-			resp, err := s.BroadcastMessages(
-				ctx,
-				chainA,
-				controllerAccount,
-				submitMsg,
-			)
-
-			s.AssertValidTxResponse(resp)
-			s.Require().NoError(err)
-		})
-
-		t.Run("verify balance is the same", func(t *testing.T) {
-			balance, err := chainB.GetBalance(ctx, chainBAccount.Bech32Address(chainB.Config().Bech32Prefix), chainB.Config().Denom)
-			s.Require().NoError(err)
-
-			expected := testvalues.StartingTokenAmount
-			s.Require().Equal(expected, balance)
-		})
-	})
-}
-
-func (s *InterchainAccountsTestSuite) TestMsgSubmitTx_SuccessfulBankSend_Incentivized() {
+func (s *IncentivizedInterchainAccountsTestSuite) TestMsgSubmitTx_SuccessfulBankSend_Incentivized() {
 	t := s.T()
 	ctx := context.TODO()
 
@@ -405,7 +193,7 @@ func (s *InterchainAccountsTestSuite) TestMsgSubmitTx_SuccessfulBankSend_Incenti
 	})
 }
 
-func (s *InterchainAccountsTestSuite) TestMsgSubmitTx_FailedBankSend_Incentivized() {
+func (s *IncentivizedInterchainAccountsTestSuite) TestMsgSubmitTx_FailedBankSend_Incentivized() {
 	t := s.T()
 	ctx := context.TODO()
